Split Road code into its own module
